Add support for non-enum-discriminated unions
This commit changes the object-based `switches` config for unions to an
array-of—arrays-based config, to allow for non-string “keys”.  This
allows for unions that use number values for the disciminant, as
opposed to the name of an enum member.

In addition, this commit makes it such that the enum-based helper
methods are only defined on enum-based unions.